\name{dna_reverse_complement_string}
\alias{dna_reverse_complement_string}
\title{dna_reverse_complement_string}
\usage{
  dna_reverse_complement_string(x)
}
\description{
  For a unique string of DNA, it returns the reversed complement sequence.
}
\examples{
  dna_reverse_complement_string(x="ATGAC")
}
