\name{dna_reverse_complement_vector}
\alias{dna_reverse_complement_vector}
\title{dna_reverse_complement_vector}
\usage{
  dna_reverse_complement_vector(x)
}
\description{
  It returns the vector with the complement bases (A-C ; G-T) and reversed.
}
\examples{
  dna_reverse_complement_vector(x=c("A","T","G","A","C"))
}
